The basic idea is following ira_lab, init commit
this ros2 pkg will try to get pcl from front and back lidar, then convert it to laserscan. The link of tfs need to be adjusted with the robots description. It builds, but still under work. Actually I was trying to port ira_laser_tools but couldn't find a way to parse topic and handle the params dynamically....